Check if optimization option is set for openmp
Without any optimization option, the compiler’s goal is to reduce the
cost of compilation and to make debugging produce the expected results.
Statements are independent:

Most optimizations are only enabled if an -O level is set on the command
line. Otherwise they are disabled, even if individual optimization flags
are specified.

The openmp-vectorize.patch enables vectorization if -fopenmp flag is
enable, however it also has to check that at least -O2 is set.

Documentation ofr GCC optimization levels:

    https://gcc.gnu.org/onlinedocs/gcc/Optimize-Options.html

Signed-off-by: Victor Rodriguez <victor.rodriguez.bahena@intel.com>