Refresh `update` target for new upstream repo
The GCC project recently switched to using git instead of SVN, and the
git mirror from SVN has been replaced with a brand new repo created by
reposurgeon. It lives at git://gcc.gnu.org/git/gcc.git.

Tag and branch naming conventions were altered as well, so refresh the
`update` target to work with the new git repo.

Note that I changed the REVISION value to use `git describe` output
instead for easy identification of the commit we source from.

(Also note that the branch name listed by `git describe` will continue
to list the previous name (gcc_9_2_0_release) for now until a new 9.2.x
stable release is made. Whether or not this was a bug with the creation
of the new git repo, or intentional, we will live with it and hope that
the upstream repo won't be rewritten again.)

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>