Fix `update` target to skip updates without code changes
The intention is to skip the package update when there are no code
changes, but the git shortlog at the beginning of the patch meant that
running `git diff` was no longer sufficient to detect when to short
circuit the process.

Fix the detection of patch changes by diffing the "current" and "new"
versions of the patch without the shortlog prepended.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>