Unify changes from libgfortran-avx package
For a long time Clear Linux has had two gcc packages. This has caused
the occasional issue with the packages not being in sync and generally
making things harder to maintain. This change moves the avx2 and
avx512 library versions into the main gcc package so that the
libgfortran-avx package can be dropped.