Move hwcaps files to specific subpackages
Move the hwcaps files to the packages of the libs they match to avoid
issues where the elf-replace tool deletes hwcap content because it isn't
currently co-installed with the base libs.

Also fixup the gccjit to have the headers and libs in the correct
place (not in dev32) and combine the go and go-lib subpackages as the
split didn't seem to have any use.

Signed-off-by: William Douglas <william.douglas@intel.com>